pull hybrid class to plugin file and update plugin name
